Revert "fix : 최상위 루트에서 access를 무분별하게 요청하는 버그 해결"
