+++ src/interceptors/LoginInterceptor.js
@@ -3,6 +3,7 @@ import axios from 'axios';
 // Axios 인스턴스 생성
 const api = axios.create({
   baseURL: import.meta.env.VITE_API_BASE_URL,
+  withCredentials: true, // 쿠키를 함께 전송
 });
 
 // 응답 인터셉터: Access Token 만료 시 처리
